lesson2: changed the class name, the code was fully changed from the viewpoint of the logical approach.
